restrict session usage to auth_proxy
