+++ pkg/api/common_test.go
@@ -5,7 +5,6 @@ import (
 	"net/http/httptest"
 	"path/filepath"
 
-	"github.com/go-macaron/session"
 	"github.com/grafana/grafana/pkg/bus"
 	"github.com/grafana/grafana/pkg/middleware"
 	m "github.com/grafana/grafana/pkg/models"
@@ -126,7 +125,6 @@ func setupScenarioContext(url string) *scenarioContext {
 
 	sc.userAuthTokenService = newFakeUserAuthTokenService()
 	sc.m.Use(middleware.GetContextHandler(sc.userAuthTokenService))
-	sc.m.Use(middleware.Sessioner(&session.Options{}, 0))
 
 	return sc
 }
