+++ pkg/middleware/session.go
@@ -1,21 +1,13 @@
 package middleware
 
-import (
-	ms "github.com/go-macaron/session"
-	"gopkg.in/macaron.v1"
+// func Sessioner(options *ms.Options, sessionConnMaxLifetime int64) macaron.Handler {
+// 	session.Init(options, sessionConnMaxLifetime)
 
-	m "github.com/grafana/grafana/pkg/models"
-	"github.com/grafana/grafana/pkg/services/session"
-)
+// 	return func(ctx *m.ReqContext) {
+// 		ctx.Next()
 
-func Sessioner(options *ms.Options, sessionConnMaxLifetime int64) macaron.Handler {
-	session.Init(options, sessionConnMaxLifetime)
-
-	return func(ctx *m.ReqContext) {
-		ctx.Next()
-
-		if err := ctx.Session.Release(); err != nil {
-			panic("session(release): " + err.Error())
-		}
-	}
-}
+// 		if err := ctx.Session.Release(); err != nil {
+// 			panic("session(release): " + err.Error())
+// 		}
+// 	}
+// }
